feat: refactor server tools function
- look to use within a stdio and streamable http interface
